requirements.txt.txt
streamlit
arabic-reshaper
python-bidi